refactor: remove unneeded server endpoint route
